hotfix: 공지사항 작성 api role check 로직 추가

+++ src/controllers/notice.controller.ts
@@ -2,15 +2,16 @@ import { RequestHandler } from "express";
 import NoticeService from "../services/notice.service";
 import { StatusCodes } from "http-status-codes";
 import logger from "../utils/logger";
+import { ApiError } from "../utils/error";
 
 export default class NoticeController {
   public getAllNotices: RequestHandler = async (req, res, next) => {
     try {
       const page = parseInt(req.query.page as string, 10) || 1;
       const size = parseInt(req.query.size as string, 10) || 10;
-    
+
       const result = await new NoticeService().getAllNotices(page, size);
-    
+
       res.status(StatusCodes.OK).json({ code: StatusCodes.OK, result });
     } catch (e: any) {
       logger.error(`${e}`);
@@ -20,8 +21,15 @@ export default class NoticeController {
 
   public createNotice: RequestHandler = async (req, res, next) => {
     try {
-      const { title, content, authorId } = req.body;
-      await new NoticeService().createNotice({ title, content, authorId });
+      if (req.role !== "ADMIN")
+        throw new ApiError("Unauthorized", StatusCodes.UNAUTHORIZED);
+
+      const { title, content } = req.body;
+      await new NoticeService().createNotice({
+        title,
+        content,
+        username: req.username,
+      });
 
       res.status(StatusCodes.CREATED).json({ code: StatusCodes.CREATED });
     } catch (e: any) {
@@ -32,6 +40,9 @@ export default class NoticeController {
 
   public deleteNotice: RequestHandler = async (req, res, next) => {
     try {
+      if (req.role !== "ADMIN")
+        throw new ApiError("Unauthorized", StatusCodes.UNAUTHORIZED);
+
       const noticeId = parseInt(req.params.noticeId, 10);
 
       await new NoticeService().deleteNotice(noticeId);
@@ -45,6 +56,9 @@ export default class NoticeController {
 
   public updateNotice: RequestHandler = async (req, res, next) => {
     try {
+      if (req.role !== "ADMIN")
+        throw new ApiError("Unauthorized", StatusCodes.UNAUTHORIZED);
+
       const noticeId = parseInt(req.params.noticeId, 10);
       const { title, content } = req.body;
